# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "Personal web",
        "description": "My portfolio",
        "contact": {
            "email": "el.ally741@gmail.com"
        },
        "license": {
            "name": "ISC",
            "url": "https://github.com/Ally4/backend-portfolio.git"
        }
    },
    "host": "allyportfolio.herokuapp.com",
    "basePath": "/",
    "schemes": ["HTTPS", "HTTP"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
    "/posts/comment/{id}": {
        "post": {
            "tags": ["Comment"],
            "summary": "Post",
            "description": "Post a comment",
            "operationId": "comment on post id",
            "parameters": [
            {
                "in": "path",
                "name": "id",
                "required": true
            },
            {
            "in": "body",
            "name": "body",
            "description": "A owner comment the post",
            "type": "array",
            "required": true,
            "schema": {
                "$ref": "#/definitions/comment"
            }
        }
        ],
            "responses": {
                "200": {
                    "description": "The comment posted successfully"
                },
                "404": {
                    "description": "Post was not found"
                }
            }
        }
    },
    "/posts/like/{id}": {
        "post": {
            "tags": ["Like"],
            "summary": "Post",
            "description": "Like a post",
            "operationId": "like a post by id",
            "parameters": [
            {
                "in": "path",
                "name": "id",
                "required": true
            }
        ],
            "responses": {
                "200": {
                    "description": "The comment posted successfully"
                },
                "404": {
                    "description": "Post was not found"
                }
            }
        }
    }
},
    "definitions": {
...
        "comment": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "required": true
                },
                "comment": {
                    "type": "string",
                    "required": true
                }
            }
        },
...
    }
}